pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clockcontrol"
version = "2.0.0"
description = "Sistema de control de asistencia para SEGIP - Extraccion de marcajes de relojes biometricos ZKTeco"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "SEGIP", email = "desarrollo@segip.gob.bo"},
    {name = "Carlos Pacha Cordova"},
]
keywords = ["zkteco", "biometrico", "asistencia", "postgresql", "segip"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "psycopg2-binary>=2.8",
    "pyzk>=0.9",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[project.scripts]
clockcontrol = "clockcontrol.cli:main"

[project.urls]
Homepage = "https://gitlab.segip.gob.bo/carlos.pacha/clockcontrol"
Repository = "https://gitlab.segip.gob.bo/carlos.pacha/clockcontrol.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["clockcontrol*"]
exclude = ["tests*", "venv*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = ["venv", ".venv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["clockcontrol"]
omit = ["tests/*", "venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
